requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
discord.py==1.1.1
idna==2.8
multidict==4.5.2
requests==2.22.0
untangle==1.1.1
urllib3==1.25.3
websockets==6.0
yarl==1.3.0
-e git://github.com/basswaver/ifunny.git#egg=ifunny
